+++ CHANGELOG.md
@@ -71,6 +71,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Generalize time series simulation time check from `PrimaryServiceWorker` [#1648](https://github.com/ie3-institute/simona/issues/1648)
 - Harmonize test resource retrieval [#1654](https://github.com/ie3-institute/simona/issues/1654)
 - Secondary service registration for EmAgent [#1651](https://github.com/ie3-institute/simona/issues/1651)
+- Generalize data time types for service registration [#1668](https://github.com/ie3-institute/simona/issues/1668)
 - Refactored `GridAgentData` [#1665](https://github.com/ie3-institute/simona/issues/1665)
 
 ### Fixed